requirements.txt
asgiref==3.7.2
black==23.11.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
Django==4.2.7
factory-boy==3.3.0
Faker==21.0.0
freezegun==1.2.2
googlemaps==4.10.0
idna==3.6
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
platformdirs==4.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
urllib3==2.1.0
